neonvm-runner: Move net & disks utils to separate files (#1185)
main.go is huge, and this seems like a reasonable place to make a cut.
